requirements.txt
alembic==1.16.1
beautifulsoup4>=4.12.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gitignore_parser==0.1.12
gunicorn==23.0.0
httpx>=0.27.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
Pillow>=10.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
Authlib==1.3.2
requests==2.31.0
email-validator==2.2.0
Markdown==3.6
openpyxl==3.1.5
